refactor: article store to article repository
